+++ Sources/Client/DialogBox_SellOrRepair.cpp
@@ -27,11 +27,9 @@ void DialogBox_SellOrRepair::on_draw()
 {
 	if (!m_game->ensure_item_configs_loaded()) return;
 	short sX, sY;
-	uint32_t time = m_game->m_cur_time;
 	std::string txt;
 
 	int item_id;
-	char item_color;
 
 	sX = m_x;
 	sY = m_y;
@@ -44,17 +42,10 @@ void DialogBox_SellOrRepair::on_draw()
 
 		item_id = m_item_index;
 
-		item_color = player().m_item_list[item_id]->m_instance.item_color;
 		{
 			CItem* sell_cfg = m_game->get_item_config(player().m_item_list[item_id]->m_id_num);
 			auto sell_draw = m_game->get_item_draw(sell_cfg ? sell_cfg->m_display_id : 0, item_atlas::pack, sell_cfg ? sell_cfg->sprite_is_female() : false);
-			if (item_color == 0)
-				sell_draw.sprite->draw(sX + 62 + 15, sY + 84 + 30, sell_draw.frame);
-			else
-			{
-				const auto& sell_tint = m_game->m_color_palette[item_color];
-				sell_draw.sprite->draw(sX + 62 + 15, sY + 84 + 30, sell_draw.frame, hb::shared::sprite::DrawParams::tint(sell_tint.r, sell_tint.g, sell_tint.b));
-			}
+			m_game->draw_item_sprite(sell_draw, sX + 62 + 15, sY + 84 + 30, player().m_item_list[item_id]->m_instance.item_color, sell_cfg);
 		}
 
 		auto itemInfo = item_name_formatter::get().format(player().m_item_list[item_id].get());
@@ -92,17 +83,10 @@ void DialogBox_SellOrRepair::on_draw()
 		draw_new_dialog_box(InterfaceNdGame2, sX, sY, 2);
 		draw_new_dialog_box(InterfaceNdText, sX, sY, 10);
 		item_id = m_item_index;
-		item_color = player().m_item_list[item_id]->m_instance.item_color;
 		{
 			CItem* rep_cfg = m_game->get_item_config(player().m_item_list[item_id]->m_id_num);
 			auto rep_draw = m_game->get_item_draw(rep_cfg ? rep_cfg->m_display_id : 0, item_atlas::pack, rep_cfg ? rep_cfg->sprite_is_female() : false);
-			if (item_color == 0)
-				rep_draw.sprite->draw(sX + 62 + 15, sY + 84 + 30, rep_draw.frame);
-			else
-			{
-				const auto& rep_tint = m_game->m_color_palette[item_color];
-				rep_draw.sprite->draw(sX + 62 + 15, sY + 84 + 30, rep_draw.frame, hb::shared::sprite::DrawParams::tint(rep_tint.r, rep_tint.g, rep_tint.b));
-			}
+			m_game->draw_item_sprite(rep_draw, sX + 62 + 15, sY + 84 + 30, player().m_item_list[item_id]->m_instance.item_color, rep_cfg);
 		}
 		auto itemInfo2 = item_name_formatter::get().format(player().m_item_list[item_id].get());
 		txt = itemInfo2.name.c_str();